Move QueryStateStoreCommitterLogs to clients module
